+++ libs/redpanda_qt_utils/qt_utils/utils.cpp
@@ -656,3 +656,8 @@ bool isInFolder(const QString &folderpath, const QString &filepath)
     QFileInfo fileInfo(filepath);
     return fileInfo.absoluteFilePath().startsWith(includeTrailingPathDelimiter(folder.absolutePath()));
 }
+
+void createFile(const QString &fileName)
+{
+    stringToFile("",fileName);
+}