+++ libs/redpanda_qt_utils/qt_utils/utils.h
@@ -106,6 +106,7 @@ QByteArray readFileToByteArray(const QString& fileName);
 
 void stringsToFile(const QStringList& list, const QString& fileName);
 void stringToFile(const QString& str, const QString& fileName);
+void createFile(const QString& fileName);
 
 /* File I/O utils */
 bool fileExists(const QString& file);